refactor(ka): remove legacy SWM migration bridge
