+++ packages/agent/src/dkg-agent-publish.ts
@@ -160,7 +160,7 @@ import {
   type QueryRequest, type QueryResponse, type QueryAccessConfig, type LookupType,
 } from '@origintrail-official/dkg-query';
 import { DKGAgentWallet, type AgentWallet } from './agent-wallet.js';
-import { sharedMemoryScopeForFinalizedLifecycle } from './finalized-lifecycle-swm.js';
+import { sharedMemoryScopeForFinalizedLifecycle } from './finalized-lifecycle-scope.js';
 
 import { ProfileManager } from './profile-manager.js';
 import { DiscoveryClient, type SkillSearchOptions, type DiscoveredAgent, type DiscoveredOffering } from './discovery.js';
