+++ packages/publisher/src/dkg-publisher.ts
@@ -1,10 +1,10 @@
-import type { NamedKnowledgeAssetGraphIdentity, Quad, SharedMemoryGraphScope, TripleStore } from '@origintrail-official/dkg-storage';
+import type { Quad, SharedMemoryGraphScope, TripleStore } from '@origintrail-official/dkg-storage';
 import type { ChainAdapter, OnChainPublishResult, AddBatchToContextGraphParams } from '@origintrail-official/dkg-chain';
 import { enrichEvmError } from '@origintrail-official/dkg-chain';
 import type { EventBus, GraphKnowledgeAssetScope, OperationContext } from '@origintrail-official/dkg-core';
 import type { AssertionSeal } from '@origintrail-official/dkg-core';
 import { DKGEvent, Logger, createOperationContext, sha256, encodeWorkspacePublishRequest, encodeEncryptedWorkspacePayload, encryptWorkspacePayload, contextGraphDataUri, contextGraphDataGraphUri, contextGraphMetaUri, contextGraphAssertionUri, contextGraphLayerUri, MemoryLayer, assertionLifecycleUri, contextGraphSubGraphUri, contextGraphSubGraphMetaUri, SYSTEM_CONTEXT_GRAPHS, validateSubGraphName, isSafeIri, assertSafeIri, assertSafeRdfTerm, assertQuadLiteralsMutf8Safe, DKG_GOSSIP_MAX_MESSAGE_BYTES, SwmGossipPayloadTooLargeError, STORAGE_ACK_MAX_STAGING_BYTES, type Ed25519Keypair, buildAuthorAttestationTypedData, buildUpdateAuthorAttestationTypedData, AUTHOR_SCHEME_VERSION_V1, TrustLevel, TRUST_LEVEL_PREDICATE, assertNoUserAuthoredTrustLevelQuads, buildTrustLevelQuads, isTrustLevelQuad, isSwmMerkleExcludedQuad, WORKSPACE_OWNER_PREDICATE, DKG_ENTITY, DKG_ROOT_ENTITY_LEGACY, ENTITY_PRED_ALT, parseAssertionSealQuads, ASSERTION_SEAL_PREDICATES, DKG_ONTOLOGY, GRAPH_KA_CONTENT_SCOPE_VERSION, LegacyKnowledgeAssetReadOnlyError, createGraphKnowledgeAssetScope, knowledgeAssetLayerGraphUri } from '@origintrail-official/dkg-core';
-import { GraphManager, PrivateContentStore, loadSharedMemoryQuadsForScope, loadSelectedSharedMemoryQuads, migrateSharedMemoryRootClosureToNamedLifecycle, resolveSharedMemoryScopeGraphs, tryReplaceGraphAtomically, type NamedLifecycleSharedMemoryMigrationResult } from '@origintrail-official/dkg-storage';
+import { GraphManager, PrivateContentStore, loadSharedMemoryQuadsForScope, loadSelectedSharedMemoryQuads, resolveSharedMemoryScopeGraphs, tryReplaceGraphAtomically } from '@origintrail-official/dkg-storage';
 import { DEFAULT_PUBLISH_EPOCHS, MAX_PUBLISH_EPOCHS, type Publisher, type PublishOptions, type PublishResult, type KAManifestEntry, type PhaseCallback, type V10CoreNodeACK, type V10ACKProviderParams, type V10ACKProviderObject, type LegacyV10ACKProvider } from './publisher.js';
 import { assertNoUserAuthoredKnowledgeAssetSkolemTerms, skolemizeByEntity, skolemizeKnowledgeAsset, skolemizeKnowledgeAssetParts } from './auto-partition.js';
 import { assertNoKnowledgeAssetPayloadNamedGraphs } from './knowledge-asset-graph-policy.js';
@@ -181,21 +181,6 @@ async function listGraphFamily(store: TripleStore, rootGraph: string): Promise<s
 type PublishedSwmCleanupPlan =
   | { kind: 'complete-family'; dataGraphs: string[] }
   | { kind: 'named-lifecycle'; dataGraphs: string[] };
-
-export interface FinalizedLifecycleSwmDescriptor {
-  /** Assertion metadata identity; may differ from the signing author. */
-  lifecycle: {
-    contextGraphId: string;
-    assertionName: string;
-    assertionLifecycleAgentAddress: string;
-    subGraphName?: string;
-  };
-  /** Canonical named-SWM identity derived from the validated finalized seal. */
-  sealAuthorScope: NamedKnowledgeAssetGraphIdentity;
-  rootEntities: readonly string[];
-}
-
-export type EnsuredFinalizedLifecycleSwmPlacement = NamedLifecycleSharedMemoryMigrationResult;
 
 /**
  * Minimal structural view of the OT-RFC-43 Option-1 KA-number allocator the
@@ -6105,89 +6090,6 @@ export class DKGPublisher implements Publisher {
    */
   private async swmGraphsUnder(bucketGraph: string): Promise<string[]> {
     return listGraphFamily(this.store, bucketGraph);
-  }
-
-  private async repairFinalizedLifecycleSwmPointer(
-    descriptor: FinalizedLifecycleSwmDescriptor,
-    targetGraph: string,
-  ): Promise<string> {
-    const { lifecycle } = descriptor;
-    const lifecycleUri = assertionLifecycleUri(
-      lifecycle.contextGraphId,
-      lifecycle.assertionLifecycleAgentAddress,
-      lifecycle.assertionName,
-      lifecycle.subGraphName,
-    );
-    const metaGraph = contextGraphMetaUri(lifecycle.contextGraphId);
-    const assertionGraphPredicate = 'http://dkg.io/ontology/assertionGraph';
-    await this.store.deleteByPattern({
-      graph: metaGraph,
-      subject: lifecycleUri,
-      predicate: assertionGraphPredicate,
-    });
-    await this.store.insert([{
-      graph: metaGraph,
-      subject: lifecycleUri,
-      predicate: assertionGraphPredicate,
-      object: targetGraph,
-    }]);
-    await this.store.flush?.();
-    return lifecycleUri;
-  }
-
-  /**
-   * Compatibility bridge for assets shared with `skipSeal` before a KA number
-   * existed. Those roots were written to the legacy SWM bucket; once
-   * `finalize(layer:"swm")` allocates a packed KA id, exact lifecycle reads
-   * must find the same payload under its canonical named graph.
-   *
-   * Insert-before-delete makes a failed migration retryable without losing the
-   * only copy. The lifecycle pointer is updated only after the data move, and a
-   * retry also repairs that pointer when the target already contains the data.
-   */
-  async ensureFinalizedLifecycleSwmPlacement(
-    descriptor: FinalizedLifecycleSwmDescriptor,
-    ctx: OperationContext,
-  ): Promise<EnsuredFinalizedLifecycleSwmPlacement> {
-    const { lifecycle, sealAuthorScope } = descriptor;
-    const { contextGraphId, subGraphName } = lifecycle;
-    const roots = [...new Set(descriptor.rootEntities)];
-    if (roots.length === 0 || roots.some((root) => !isSafeIri(root))) {
-      throw new Error('Named lifecycle SWM migration requires at least one safe root entity IRI');
-    }
-
-    const swmGraph = this.graphManager.sharedMemoryUri(contextGraphId, subGraphName);
-    const migration = await migrateSharedMemoryRootClosureToNamedLifecycle(
-      this.store,
-      swmGraph,
-      { rootEntities: roots },
-      sealAuthorScope,
-      {
-        graphResolution: {
-          source: 'publisher.ensureFinalizedLifecycleSwmPlacement.graphs',
-        },
-        namedSource: {
-          source: 'publisher.ensureFinalizedLifecycleSwmPlacement.named-source',
-        },
-        legacyBucketSource: {
-          source: 'publisher.ensureFinalizedLifecycleSwmPlacement.legacy-source',
-        },
-      },
-    );
-    if (migration.sourceEmpty) return migration;
-
-    const lifecycleUri = await this.repairFinalizedLifecycleSwmPointer(
-      descriptor,
-      migration.targetGraph,
-    );
-    if (migration.migratedLegacyQuadCount > 0) {
-      this.log.info(
-        ctx,
-        `Copied ${migration.migratedLegacyQuadCount} legacy SWM triple(s) ` +
-        `for <${lifecycleUri}> to <${migration.targetGraph}> while retaining the compatibility source`,
-      );
-    }
-    return migration;
   }
 
   /**
