+++ packages/publisher/test/shared-memory-publish-boundary.test.ts
@@ -9,7 +9,6 @@ import {
   generateEd25519Keypair,
   DKG_ENTITY,
   DKG_ROOT_ENTITY_LEGACY,
-  assertionLifecycleUri,
 } from '@origintrail-official/dkg-core';
 import { OxigraphStore, type Quad } from '@origintrail-official/dkg-storage';
 import {
@@ -238,249 +237,6 @@ describe('publishFromSharedMemory multi-root selection (OT-RFC-44 / Design B: on
     ]);
   });
 
-  it('migrates a finalized skipSeal payload from the legacy bucket into its canonical named graph', async () => {
-    const { publisher, store } = await makeRealPublisher();
-    const querySources: Array<string | undefined> = [];
-    const realQuery = store.query.bind(store);
-    store.query = async (sparql, options) => {
-      querySources.push(options?.source);
-      return realQuery(sparql, options);
-    };
-    const root = 'urn:test:root:skip-seal';
-    const child = `${root}/.well-known/genid/child`;
-    const checksummedAuthor = '0xAbCdEf0123456789AbCdEf0123456789AbCdEf01';
-    const canonicalGraph = `${SWM_GRAPH}/${checksummedAuthor.toLowerCase()}/7`;
-    const lifecycle = assertionLifecycleUri(CONTEXT_GRAPH, checksummedAuthor, 'skip-seal');
-    await store.insert([
-      q(root, 'http://schema.org/name', '"legacy"', SWM_GRAPH),
-      q(child, 'http://schema.org/value', '"child"', SWM_GRAPH),
-      {
-        subject: lifecycle,
-        predicate: 'http://dkg.io/ontology/assertionGraph',
-        object: SWM_GRAPH,
-        graph: `${CONTEXT_GRAPH_URI}/_meta`,
-      },
-    ]);
-    const ensured = await publisher.ensureFinalizedLifecycleSwmPlacement(
-      {
-        lifecycle: {
-          contextGraphId: CONTEXT_GRAPH,
-          assertionName: 'skip-seal',
-          assertionLifecycleAgentAddress: checksummedAuthor,
-        },
-        sealAuthorScope: { agentAddress: checksummedAuthor, kaNumber: 7n },
-        rootEntities: [root],
-      },
-      createOperationContext('test'),
-    );
-
-    expect(ensured.sourceEmpty).toBe(false);
-    expect(ensured.migratedLegacyQuadCount).toBe(2);
-    expect(ensured.targetGraph).toBe(canonicalGraph);
-    expect(ensured).not.toHaveProperty('quads');
-    expect(querySources.some((source) => source?.endsWith('.result'))).toBe(false);
-    const legacyRoot = await store.query(
-      `ASK { GRAPH <${SWM_GRAPH}> { <${root}> ?p ?o } }`,
-    );
-    const legacyChild = await store.query(
-      `ASK { GRAPH <${SWM_GRAPH}> { <${child}> ?p ?o } }`,
-    );
-    expect(legacyRoot).toMatchObject({ type: 'boolean', value: true });
-    expect(legacyChild).toMatchObject({ type: 'boolean', value: true });
-    expect(await store.deleteByPattern({ graph: canonicalGraph, subject: root })).toBe(1);
-    expect(await store.deleteByPattern({ graph: canonicalGraph, subject: child })).toBe(1);
-    const pointer = await store.query(
-      `SELECT ?g WHERE { GRAPH <${CONTEXT_GRAPH_URI}/_meta> { ` +
-      `<${lifecycle}> <http://dkg.io/ontology/assertionGraph> ?g } }`,
-    );
-    expect(pointer.type).toBe('bindings');
-    expect(pointer.type === 'bindings' ? pointer.bindings[0]?.['g'] : undefined).toBe(canonicalGraph);
-  });
-
-  it('preserves a same-root legacy copy so a sibling can migrate after the first lifecycle publishes', async () => {
-    const { publisher, store, publishSpy } = await makePublisher(
-      new NoChainAdapter(),
-      'confirmed',
-    );
-    const root = 'urn:test:root:legacy-siblings';
-    const firstAuthor = '0x1111111111111111111111111111111111111111';
-    const secondAuthor = '0x2222222222222222222222222222222222222222';
-    const firstGraph = `${SWM_GRAPH}/${firstAuthor}/7`;
-    const secondGraph = `${SWM_GRAPH}/${secondAuthor}/9`;
-    await store.insert([q(root, 'http://schema.org/name', '"shared legacy root"', SWM_GRAPH)]);
-
-    const first = await publisher.ensureFinalizedLifecycleSwmPlacement(
-      {
-        lifecycle: {
-          contextGraphId: CONTEXT_GRAPH,
-          assertionName: 'legacy-sibling-a',
-          assertionLifecycleAgentAddress: firstAuthor,
-        },
-        sealAuthorScope: { agentAddress: firstAuthor, kaNumber: 7n },
-        rootEntities: [root],
-      },
-      createOperationContext('test'),
-    );
-    await publisher.publishFromSharedMemory(
-      CONTEXT_GRAPH,
-      { rootEntities: [root] },
-      {
-        sharedMemoryScope: {
-          kind: 'named-lifecycle',
-          identity: { agentAddress: firstAuthor, kaNumber: 7n },
-        },
-      },
-    );
-    const second = await publisher.ensureFinalizedLifecycleSwmPlacement(
-      {
-        lifecycle: {
-          contextGraphId: CONTEXT_GRAPH,
-          assertionName: 'legacy-sibling-b',
-          assertionLifecycleAgentAddress: secondAuthor,
-        },
-        sealAuthorScope: { agentAddress: secondAuthor, kaNumber: 9n },
-        rootEntities: [root],
-      },
-      createOperationContext('test'),
-    );
-    await publisher.publishFromSharedMemory(
-      CONTEXT_GRAPH,
-      { rootEntities: [root] },
-      {
-        sharedMemoryScope: {
-          kind: 'named-lifecycle',
-          identity: { agentAddress: secondAuthor, kaNumber: 9n },
-        },
-      },
-    );
-
-    const legacy = await store.query(`ASK { GRAPH <${SWM_GRAPH}> { <${root}> ?p ?o } }`);
-    const firstNamed = await store.query(`ASK { GRAPH <${firstGraph}> { <${root}> ?p ?o } }`);
-    const secondNamed = await store.query(`ASK { GRAPH <${secondGraph}> { <${root}> ?p ?o } }`);
-    expect(first).toMatchObject({ sourceEmpty: false, migratedLegacyQuadCount: 1 });
-    expect(second).toMatchObject({ sourceEmpty: false, migratedLegacyQuadCount: 1 });
-    expect(legacy).toMatchObject({ type: 'boolean', value: true });
-    expect(firstNamed).toMatchObject({ type: 'boolean', value: false });
-    expect(secondNamed).toMatchObject({ type: 'boolean', value: false });
-    expect(publishSpy.calls).toHaveLength(2);
-    for (const [input] of publishSpy.calls) {
-      expect(input.quads).toEqual([
-        { subject: root, predicate: 'http://schema.org/name', object: '"shared legacy root"', graph: '' },
-      ]);
-    }
-  });
-
-  it('repairs the lifecycle pointer when a retry finds only the canonical copy', async () => {
-    const { publisher, store } = await makeRealPublisher();
-    const root = 'urn:test:root:skip-seal-retry';
-    const checksummedAuthor = '0xAbCdEf0123456789AbCdEf0123456789AbCdEf01';
-    const canonicalGraph = `${SWM_GRAPH}/${checksummedAuthor.toLowerCase()}/8`;
-    const lifecycle = assertionLifecycleUri(CONTEXT_GRAPH, checksummedAuthor, 'skip-seal-retry');
-    await store.insert([
-      q(root, 'http://schema.org/name', '"canonical"', canonicalGraph),
-      {
-        subject: lifecycle,
-        predicate: 'http://dkg.io/ontology/assertionGraph',
-        object: SWM_GRAPH,
-        graph: `${CONTEXT_GRAPH_URI}/_meta`,
-      },
-    ]);
-
-    const ensured = await publisher.ensureFinalizedLifecycleSwmPlacement(
-      {
-        lifecycle: {
-          contextGraphId: CONTEXT_GRAPH,
-          assertionName: 'skip-seal-retry',
-          assertionLifecycleAgentAddress: checksummedAuthor,
-        },
-        sealAuthorScope: { agentAddress: checksummedAuthor, kaNumber: 8n },
-        rootEntities: [root],
-      },
-      createOperationContext('test'),
-    );
-
-    expect(ensured.sourceEmpty).toBe(false);
-    expect(ensured.migratedLegacyQuadCount).toBe(0);
-    expect(ensured.targetGraph).toBe(canonicalGraph);
-    const canonical = await store.query(`ASK { GRAPH <${canonicalGraph}> { <${root}> ?p ?o } }`);
-    expect(canonical).toMatchObject({ type: 'boolean', value: true });
-    const pointer = await store.query(
-      `SELECT ?g WHERE { GRAPH <${CONTEXT_GRAPH_URI}/_meta> { ` +
-      `<${lifecycle}> <http://dkg.io/ontology/assertionGraph> ?g } }`,
-    );
-    expect(pointer.type).toBe('bindings');
-    expect(pointer.type === 'bindings' ? pointer.bindings[0]?.['g'] : undefined).toBe(canonicalGraph);
-  });
-
-  it('returns an explicit source-empty result without repairing lifecycle metadata', async () => {
-    const { publisher, store } = await makeRealPublisher();
-    const root = 'urn:test:root:source-empty';
-    const author = '0xAbCdEf0123456789AbCdEf0123456789AbCdEf01';
-    const canonicalGraph = `${SWM_GRAPH}/${author.toLowerCase()}/10`;
-    const lifecycle = assertionLifecycleUri(CONTEXT_GRAPH, author, 'source-empty');
-    await store.insert([{
-      subject: lifecycle,
-      predicate: 'http://dkg.io/ontology/assertionGraph',
-      object: SWM_GRAPH,
-      graph: `${CONTEXT_GRAPH_URI}/_meta`,
-    }]);
-
-    const ensured = await publisher.ensureFinalizedLifecycleSwmPlacement(
-      {
-        lifecycle: {
-          contextGraphId: CONTEXT_GRAPH,
-          assertionName: 'source-empty',
-          assertionLifecycleAgentAddress: author,
-        },
-        sealAuthorScope: { agentAddress: author, kaNumber: 10n },
-        rootEntities: [root],
-      },
-      createOperationContext('test'),
-    );
-
-    expect(ensured).toEqual({
-      sourceEmpty: true,
-      migratedLegacyQuadCount: 0,
-      targetGraph: canonicalGraph,
-    });
-    const pointer = await store.query(
-      `SELECT ?g WHERE { GRAPH <${CONTEXT_GRAPH_URI}/_meta> { ` +
-      `<${lifecycle}> <http://dkg.io/ontology/assertionGraph> ?g } }`,
-    );
-    expect(pointer.type === 'bindings' ? pointer.bindings[0]?.['g'] : undefined).toBe(SWM_GRAPH);
-  });
-
-  it('converges a checksum-cased alias without issuing a graphless delete', async () => {
-    const { publisher, store } = await makeRealPublisher();
-    const root = 'urn:test:root:checksum-alias';
-    const checksummedAuthor = '0xAbCdEf0123456789AbCdEf0123456789AbCdEf01';
-    const aliasGraph = `${SWM_GRAPH}/${checksummedAuthor}/9`;
-    const canonicalGraph = `${SWM_GRAPH}/${checksummedAuthor.toLowerCase()}/9`;
-    await store.insert([
-      q(root, 'http://schema.org/name', '"legacy alias"', aliasGraph),
-    ]);
-
-    const ensured = await publisher.ensureFinalizedLifecycleSwmPlacement(
-      {
-        lifecycle: {
-          contextGraphId: CONTEXT_GRAPH,
-          assertionName: 'checksum-alias',
-          assertionLifecycleAgentAddress: checksummedAuthor,
-        },
-        sealAuthorScope: { agentAddress: checksummedAuthor, kaNumber: 9n },
-        rootEntities: [root],
-      },
-      createOperationContext('test'),
-    );
-
-    expect(ensured.sourceEmpty).toBe(false);
-    expect(ensured.migratedLegacyQuadCount).toBe(0);
-    expect(ensured.targetGraph).toBe(canonicalGraph);
-    const alias = await store.query(`ASK { GRAPH <${aliasGraph}> { <${root}> ?p ?o } }`);
-    const canonical = await store.query(`ASK { GRAPH <${canonicalGraph}> { <${root}> ?p ?o } }`);
-    expect(alias).toMatchObject({ type: 'boolean', value: false });
-    expect(canonical).toMatchObject({ type: 'boolean', value: true });
-  });
 
   it('confirmed exact cleanup removes stale ownership when the local KA was the last share', async () => {
     const { publisher, store } = await makePublisher(new NoChainAdapter(), 'confirmed');
