+++ packages/storage/src/index.ts
@@ -95,7 +95,6 @@ export {
   GraphManager,
   loadSelectedSharedMemoryQuads,
   loadSharedMemoryQuadsForScope,
-  migrateSharedMemoryRootClosureToNamedLifecycle,
   loadSharedMemorySliceWithKaBoundFallback,
   canonicalSharedMemoryScopeWriteGraph,
   resolveSharedMemoryScopeGraphs,
@@ -104,8 +103,6 @@ export {
   resolveSharedMemoryReadGraphs,
   resolveVerifiableMemoryReadGraphs,
   type LoadSelectedSharedMemoryQuadsOptions,
-  type MigrateNamedLifecycleSharedMemoryOptions,
-  type NamedLifecycleSharedMemoryMigrationResult,
   type LoadSelectedVerifiableMemoryQuadsOptions,
   type NonEmptyGraphList,
   type NamedKnowledgeAssetGraphIdentity,
